Refactor module structure to clarify an event whether public or internal
