+++ src/behaviour.rs
@@ -1,7 +1,7 @@
 use crate::beacon_chain::BeaconChain;
-use crate::discovery::behaviour::DiscoveryEvent;
+use crate::discovery::DiscoveryEvent;
 use crate::peer_manager::PeerManagerEvent;
-use crate::rpc::behaviour::RpcEvent;
+use crate::rpc::RpcEvent;
 use libp2p::swarm::dial_opts::{DialOpts, PeerCondition};
 use libp2p::swarm::handler::DummyConnectionHandler;
 use libp2p::swarm::NetworkBehaviour;
